backend/requirements.txt
Django>=4.1.3<5.0
djangorestframework
djangorestframework_simplejwt
django_rest_client
django-encrypted-model-fields
django-storages
django-filter
django-environ
channels[daphne]>=4.0.0
djangochannelsrestframework==1.1.0
channels_redis
asgi_redis
websockets
django-typomatic
psycopg2>=2.8
boto3
Mastodon.py>=1.7.0
twitchAPI>=3.2.0.1
hikari>=2.0.0.dev112
google-cloud-storage
markdown
Pillow
